+++ src/main/java/com/example/postreadservice/application/PostReadService.java
@@ -1,0 +1,11 @@
+package com.example.postreadservice.application;
+
+import com.example.postreadservice.dto.out.PostReadModelResDto;
+import com.example.postreadservice.kafka.event.PostCreatedEvent;
+
+public interface PostReadService {
+    void createPostRead(PostCreatedEvent postCreatedEvent);
+
+    PostReadModelResDto getPostRead(String postUuid);
+
+}
