+++ src/main/java/com/example/postreadservice/common/exception/BaseException.java
@@ -1,0 +1,14 @@
+package com.example.postreadservice.common.exception;
+
+import com.example.postreadservice.common.response.BaseResponseStatus;
+import lombok.Getter;
+
+@Getter
+public class BaseException extends RuntimeException {
+
+    private final BaseResponseStatus status;
+
+    public BaseException(BaseResponseStatus status) {
+        this.status = status;
+    }
+}
