+++ src/main/java/com/example/postreadservice/common/exception/BaseExceptionHandler.java
@@ -1,0 +1,35 @@
+package com.example.postreadservice.common.exception;
+
+import com.example.postreadservice.common.response.BaseResponseStatus;
+import com.example.postreadservice.common.entity.BaseResponseEntity;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+@RestControllerAdvice
+@Slf4j
+public class BaseExceptionHandler {
+
+    /**
+     * 발생한 예외 처리
+     */
+
+    @ExceptionHandler(BaseException.class)
+    protected ResponseEntity<BaseResponseEntity<Void>> BaseError(BaseException e) {
+        BaseResponseEntity<Void> response = new BaseResponseEntity<>(e.getStatus());
+        log.error("BaseException -> {}({})", e.getStatus(), e.getStatus().getMessage(), e);
+        return new ResponseEntity<>(response, response.httpStatus());
+    }
+
+    @ExceptionHandler(RuntimeException.class)
+    protected ResponseEntity<BaseResponseEntity<Void>> RuntimeError(RuntimeException e) {
+        BaseResponseEntity<Void> response = new BaseResponseEntity<>(BaseResponseStatus.INTERNAL_SERVER_ERROR, e.getMessage());
+        log.error("RuntimeException: ", e);
+        for (StackTraceElement s : e.getStackTrace()) {
+            System.out.println(s);
+        }
+        return new ResponseEntity<>(response, response.httpStatus());
+    }
+
+}
