+++ src/main/java/com/example/postreadservice/entity/Image.java
@@ -1,0 +1,15 @@
+package com.example.postreadservice.entity;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class Image {
+    private String imageUrl;
+    private String alt;
+}
