--- src/main/java/com/example/readservice/application/PostReadServiceImpl.java
@@ -1,38 +1,0 @@
-package com.example.readservice.application;
-
-import com.example.readservice.common.entity.BaseResponseStatus;
-import com.example.readservice.common.exception.BaseException;
-import com.example.readservice.dto.in.PostReadModelReqDto;
-import com.example.readservice.dto.out.PostReadModelResDto;
-import com.example.readservice.entity.PostReadModel;
-import com.example.readservice.infrastructure.PostReadRepository;
-import com.example.readservice.kafka.event.PostCreatedEvent;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-
-@Service
-@RequiredArgsConstructor
-@Slf4j
-@Transactional(readOnly = true)
-public class PostReadServiceImpl implements PostReadService {
-
-    private final PostReadRepository postReadRepository;
-
-    @Transactional
-    @Override
-    public void createPostRead(PostCreatedEvent postCreatedEvent) {
-        postReadRepository.save(PostReadModelReqDto.from(postCreatedEvent));
-    }
-
-    @Override
-    public PostReadModelResDto getPostRead(String postUuid) {
-        PostReadModel postReadModel = postReadRepository.findByPostUuid(postUuid)
-                .orElseThrow(() -> new BaseException(BaseResponseStatus.POST_NOT_FOUND));
-        return PostReadModelResDto.from(postReadModel);
-    }
-
-
-}
