--- src/main/java/com/example/readservice/common/config/SwaggerConfig.java
@@ -1,48 +1,0 @@
-package com.example.readservice.common.config;
-
-import io.swagger.v3.oas.annotations.OpenAPIDefinition;
-import io.swagger.v3.oas.annotations.tags.Tag;
-import io.swagger.v3.oas.models.Components;
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Info;
-import io.swagger.v3.oas.models.security.SecurityRequirement;
-import io.swagger.v3.oas.models.security.SecurityScheme;
-import io.swagger.v3.oas.models.servers.Server;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@OpenAPIDefinition(
-        tags = {
-                @Tag(name = "post-read", description = "질문 조회 관련 API 입니다.")
-        }
-)
-@Configuration
-public class SwaggerConfig {
-
-    private static final String BEARER_TOKEN_PREFIX = "Bearer";
-
-    @Bean
-    public OpenAPI openAPI() {
-        String securityJwtName = "JWT";
-        SecurityRequirement securityRequirement = new SecurityRequirement().addList(securityJwtName);
-        Components components = new Components()
-                .addSecuritySchemes(securityJwtName, new SecurityScheme()
-                        .name(securityJwtName)
-                        .type(SecurityScheme.Type.HTTP)
-                        .scheme(BEARER_TOKEN_PREFIX)
-                        .bearerFormat(securityJwtName));
-
-        return new OpenAPI()
-                .addSecurityItem(securityRequirement)
-                .components(components)
-                .addServersItem(new Server().url("/post-read-service"))
-                .info(apiInfo());
-    }
-
-    private Info apiInfo() {
-        return new Info()
-                .title("POST-READ-SERVICE API DOCS")
-                .description("post-read-service API 테스트를 위한 Swagger UI")
-                .version("1.0.0");
-    }
-}
