--- src/main/java/com/example/readservice/common/entity/BaseDocument.java
@@ -1,20 +1,0 @@
-package com.example.readservice.common.entity;
-
-import lombok.Getter;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.mongodb.core.mapping.Field;
-
-import java.time.LocalDateTime;
-
-@Getter
-public abstract class BaseDocument {
-
-    @CreatedDate
-    @Field("created_at")
-    private LocalDateTime createdAt;
-
-    @LastModifiedDate
-    @Field("updated_at")
-    private LocalDateTime updatedAt;
-}
