--- src/main/java/com/example/readservice/common/exception/BaseException.java
@@ -1,14 +1,0 @@
-package com.example.readservice.common.exception;
-
-import com.example.readservice.common.entity.BaseResponseStatus;
-import lombok.Getter;
-
-@Getter
-public class BaseException extends RuntimeException {
-
-    private final BaseResponseStatus status;
-
-    public BaseException(BaseResponseStatus status) {
-        this.status = status;
-    }
-}
