--- src/main/java/com/example/readservice/common/exception/BaseExceptionHandler.java
@@ -1,37 +1,0 @@
-package com.example.readservice.common.exception;
-
-import com.example.readservice.common.entity.BaseResponseEntity;
-import com.example.readservice.common.entity.BaseResponseStatus;
-import io.swagger.v3.oas.annotations.Hidden;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@Hidden
-@RestControllerAdvice
-@Slf4j
-public class BaseExceptionHandler {
-
-    /**
-     * 발생한 예외 처리
-     */
-
-    @ExceptionHandler(BaseException.class)
-    protected ResponseEntity<BaseResponseEntity<Void>> BaseError(BaseException e) {
-        BaseResponseEntity<Void> response = new BaseResponseEntity<>(e.getStatus());
-        log.error("BaseException -> {}({})", e.getStatus(), e.getStatus().getMessage(), e);
-        return new ResponseEntity<>(response, response.httpStatus());
-    }
-
-    @ExceptionHandler(RuntimeException.class)
-    protected ResponseEntity<BaseResponseEntity<Void>> RuntimeError(RuntimeException e) {
-        BaseResponseEntity<Void> response = new BaseResponseEntity<>(BaseResponseStatus.INTERNAL_SERVER_ERROR, e.getMessage());
-        log.error("RuntimeException: ", e);
-        for (StackTraceElement s : e.getStackTrace()) {
-            System.out.println(s);
-        }
-        return new ResponseEntity<>(response, response.httpStatus());
-    }
-
-}
