--- src/main/java/com/example/readservice/entity/Image.java
@@ -1,15 +1,0 @@
-package com.example.readservice.entity;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class Image {
-    private String imageUrl;
-    private String alt;
-}
