--- src/main/java/com/example/readservice/presentation/PostReadController.java
@@ -1,25 +1,0 @@
-package com.example.readservice.presentation;
-
-import com.example.readservice.application.PostReadService;
-import com.example.readservice.common.response.BaseResponseEntity;
-import com.example.readservice.dto.out.PostReadModelResDto;
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/api/v1/post-read")
-@RequiredArgsConstructor
-public class PostReadController {
-
-    private final PostReadService postReadService;
-
-    @GetMapping("/{postUuid}")
-    public BaseResponseEntity<PostReadModelResDto> getPostRead(@PathVariable String postUuid) {
-        return new BaseResponseEntity<>(postReadService.getPostRead(postUuid));
-    }
-
-
-}
