+++ src/main/resources/application.yml
@@ -4,7 +4,7 @@ server:
 
 spring:
   application:
-    name: read-service
+    name: post-read-service
   profiles:
     active: dev
