+++ src/test/java/com/example/postreadservice/PostreadserviceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.example.postreadservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class PostreadserviceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
