--- src/test/java/com/example/readservice/ReadserviceApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.readservice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ReadserviceApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
